memo_operationクラスを作成
